Desktop-Borna/Devops/Bots/Telegram_to_wordpress_bot/requirements.txt
certifi==2020.6.20
cffi==1.14.1
cryptography==3.0
decorator==4.4.2
pycparser==2.20
python-telegram-bot==12.8
python-wordpress-xmlrpc==2.3
six==1.15.0
tornado==6.0.4
